pyproject.toml
[tool.poetry]
name = "inline-problem"
version = "0.1.0"
description = "Problem with Inline Form"
authors = ["Gary Jarrel <gary@jarrel.com.au>"]
readme = "README.md"
packages = [{include = "inline_problem"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "4.1.3"
django-auto-prefetch = "^1.3.0"

[tool.poetry.group.dev.dependencies]
autoflake = "^1.4"
black = "^22.3.0"
coverage = "^6.3.1"
django-coverage-plugin = "^2.0.3"
django-debug-toolbar = "^3.2.4"
django-extensions = "^3.1.5"
django-stubs = { extras = ["compatible-mypy"], version = "^1.12.0" }
django-test-migrations = "^1.2.0"
factory-boy = "^3.2.1"
flake8 = "^5.0.4"
flake8-isort = "^4.1.1"
isort = "^5.10.1"
mypy = "0.961"
pre-commit = "^2.20.0"
pylint-django = "^2.5.0"
pytest = "7.2.0"
pytest-asyncio = "^0.19.0"
pytest-django = "^4.5.2"
pytest-mock = "^3.7.0"
pytest-randomly = "^3.11.0"
pytest-sugar = "^0.9.4"
requests-mock = "^1.9.3"
tbump = "^6.9.0"
types-cachetools = "^5.2.1"
types-python-dateutil = "^2.8.19"
whitenoise = "^6.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
